\name{backpay-package}
\alias{backpay-package}
\alias{backpay}
\docType{package}
\title{
\packageTitle{backpay}
}
\description{
\packageDescription{backpay}
}
%\details{
%The DESCRIPTION file:
%\packageDESCRIPTION{backpay}
%\packageIndices{backpay}
%~~ An overview of how to use the package, including the most important ~~
%~~ functions ~~
%}
\author{
\packageAuthor{backpay}

Maintainer: \packageMaintainer{backpay}
}
\references{
Thierry Chekouo et al (2018), \emph{Investigating Protein Patterns in Human Leukemia Cell Line Experiments: A Bayesian Approach for Extremely Small Sample Sizes}, submitted.
}

\keyword{ package }
\seealso{
\code{\link{BAckPay}}, \code{\link{Generate}}
}
%\examples{
%~~ simple examples of the most important functions ~~
%}
